feat(Station): 지하철 정보를 나타내는 클래스
