+++ src/main/java/subway/domain/subway/station/Station.java
@@ -1,0 +1,37 @@
+package subway.domain.subway.station;
+
+import java.util.Objects;
+
+public class Station {
+    private final String name;
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Station station = (Station) o;
+        return Objects.equals(name, station.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
+    }
+
+    @Override
+    public String toString() {
+        return "Station{" + name + "}";
+    }
+}
